refactor: 優化 NotificationList 組件
