+++ src/lib/is-client.ts
@@ -1,0 +1,3 @@
+const isClient = () => typeof window !== "undefined";
+
+export default isClient;
